# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Digi Assesment Doc",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "kopiaman@yahoo.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://digiback.test/api",
            "description": "Local Development"
        },
        {
            "url": "http://my-default-host.com",
            "description": "Live Server"
        }
    ],
    "paths": {
        "/auth/create": {
            "post": {
                "summary": "Create new user",
                "operationId": "App\\Http\\Controllers\\API\\AuthController::create",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        },
        "/auth/login": {
            "post": {
                "summary": "Login existing user",
                "operationId": "App\\Http\\Controllers\\API\\AuthController::login",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        },
        "/users": {
            "get": {
                "summary": "List of users name",
                "operationId": "App\\Http\\Controllers\\API\\UserController::index",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        }
    }
}